+++ index.js
@@ -36,6 +36,7 @@ require('./models/gateModel.js');
 app.use(require('./middlewares/flash.js'));
 app.use(require('./middlewares/verification.js'));
 app.use(require('./middlewares/username.js'));
+app.use(require('./middlewares/passSession.js'));
 /****/
 
 /*Routers*/